id utente Long

+++ src/modello/Utente.java
@@ -2,15 +2,16 @@ package modello;
 
 public class Utente {
 	
-	private int id;
+	private Long id;
 	private String nome;
 	private String cognome;
 	private String codiceFIscale;
 	private int eta;
-	public int getId() {
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNome() {
